requirements.txt
langchain==0.1.12
langchain-community==0.0.24
streamlit==1.24.0
python-dotenv==1.0.0

# PDF parsing
PyPDF2==3.0.1
pdfplumber==0.10.2

# Vector store
faiss-cpu==1.7.4

# Embeddings
sentence-transformers==2.2.2
huggingface-hub==0.14.1

# openai==0.27.6   # uncomment only if using OpenAI models

# Utilities
altair==4.2.0
tiktoken==0.4.0
